### 2025-05-27 #8287 @editorialbot
[PRE REVIEW]: RealPaver 1.1: A C++ Library for Constraint Programming over Numeric or Mixed Discrete-Continuous Domains
**Submitting author:** @realpaver (<a href="http://orcid.org/0000-0002-9761-6703">Raphael Chenouard</a>)
**Repository:** https://github.com/realpaver/realpaver
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.1.1
