### 2025-05-27 #8287 @crvernon
@realpaver - thank you for your submission to JOSS.  A few things for you to address before we get started:

- Please reduce the word count of your paper to <= 1000 words
- Your software uses the Gurobi solver which is not open source and may not be accessible to all reviewers.  Please consider providing an option for an open source solver as well so all users can utilize and test your software.

We are working through a large backlog of submission right now but will try to find you a topic editor as soon as possible.
