### 2025-05-28 #8287 @raphaelchenouard
Thank you for your message.

Regarding your demand about reducing the number of words, we'll soon make a commit with a reduced description of the library and without the example we used as an illustration.

Gurobi is only one of the compatible linear solvers that can be used. The open-source CLP solver is the one used by default, and SoPlex and Highs can also be defined at the configuration step of the building process of the library.
