### 2025-06-03 #8287 @pachadotdev
> [@editorialbot](https://github.com/editorialbot) invite [@HaoZeke](https://github.com/HaoZeke) as editor
>
> [@HaoZeke](https://github.com/HaoZeke) do you have time to take this one on as editor?

hi @crvernon happy to take this one
