### 2025-06-03 bee1a40
Update paper.md to reduce number of words
