+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 Constraint Programming (CP) is a paradigm for solving constraint satisfaction and optimization problems [@rossi2006]. Although CP mainly addresses combinatorial problems, it can also handle continuous problems by approximating real numbers with intervals [@benhamou2006]. 
 
-RealPaver is a C++ library for CP over numeric or mixed discrete-continuous domains. Constraint Satisfaction Prblems (CSPs) can be described either in C++ with the API, or in a text file using the syntax of RealPaver specific language.  Then, they can be solved using the C++ API or using the CSP solver from the command-line. The CSP solver is pre-configured, but various parameters can be modified in another text file.
+RealPaver is a C++ library for CP over numeric or mixed discrete-continuous domains. Constraint Satisfaction Problems (CSPs) can be described either in C++ with the API or in a text file using the syntax of RealPaver specific language.  Then, they can be solved using the C++ API or using the CSP solver from the command line. The CSP solver is pre-configured, but various parameters can be modified in another text file.
 
 With respect to the first version of the software developed twenty years ago [@granvilliers2006], this new library incorporates new types of variables and constraints, new algorithms, a clean object-oriented architecture, the management of parameters, Meson Build as build engine [@meson], an interface with third-party softwares and a C++ API. It achieves performances equivalent to the competing library Ibex [@ibex] for pure continuous problems.
 
@@ -38,13 +38,11 @@ CP associates a rich modeling language with powerful solving techniques. The mai
 
 This technology has been applied with success in many fields of engineering like automatic control [@jaulin2001], preliminary design [@yvars2021] and robotics [@merlet2004].
 
-This library can be used by anyone wanting to compute sets of solutions for numerical or mixed discrete-continuous constraint satisfaction problems. It can also be used to prove infeasibility or existence of solutions thanks to the B\&P algorithm and interval analysis [@moore2009].
+This library can be used by anyone wanting to compute sets of solutions for numerical or mixed discrete-continuous constraint satisfaction problems. It can also be used to prove infeasibility or existence of solutions, thanks to the B\&P algorithm and interval analysis [@moore2009].
 Since the library contains most of the state-of-the art algorithms relating to CP over intervals, it can also be used by researchers in this field to define new algorithms.
 
 
-# Overview
-
-The RealPaver libray provides a CSP solver relying on a modern C++ API. Problems can be defined through the object-oriented API or using a domain specific language.
+# Brief overview
 
 ## Modeling language
 
@@ -57,25 +55,21 @@ There are three types of variables:
 The language defines several types of constraints:
 
 * An arithmetic constraint involves the usual operations over the reals and relations from the set $\{=, \leq, \geq\}$;
-* $G\to B$ is a conditional constraint where a constraint $B$ (body) is activated when a constraint $G$ (guard) holds true;
-* $table(X, S)$ is a table constraint where $X$ is a vector of variables and $S$ is a set of valid assignments for $X$;
+* $G\to B$ is a conditional constraint, where a constraint $B$ (body) is activated when a constraint $G$ (guard) holds true;
+* $table(X, S)$ is a table constraint, where $X$ is a vector of variables and $S$ is a set of valid assignments for $X$;
 * $piecewise(x, \{I_k\to C_k\}_{k})$ is a piecewise constraint where a constraint $C_k$ is activated when the variable $x$ lies in the interval $I_k$.
 
-## Search strategies
+## Solving strategies
 
[…]
