### 2025-06-03 #8287 @danielskatz
👋 @pachadotdev - the command you wanted is the one @crvernon used above, or more simply `@editorialbot assign me as editor`
