### 2025-06-16 #8287 @raphaelchenouard
Hi, do I have something else to do on my side? Is it ok if the number of words in the paper is 1015 (I used the lua script available on pandoc and it was more around 900) ?
