### 2025-06-16 #8287 @raphaelchenouard
Here is a list of a few persons I identified:
- cvanaret
- victorapm
- lukeolson
- dannys4
- alberto743
