### 2025-08-06 #8287 @alberto743
Ho @pachadotdev.
Interesting work. Unfortunately I am quite busy at the moment, and I guess that I will not be able to perform this review right now, given also the amount of lines of code involved.
May I suggest @sbna, instead?
