### 2025-09-22 #8287 @pachadotdev
> Hi [@pachadotdev](https://github.com/pachadotdev), I'd be interested. When would you need the review and how extensive should it be?

hi @crvernon, thanks a lot
it needs to be a brief review of the paper, nothing like a monograph
if that's ok, I'll add you as a reviewer
