### 2025-09-22 #8287 @crvernon
> > Hi [@pachadotdev](https://github.com/pachadotdev), I'd be interested. When would you need the review and how extensive should it be?
>
> hi [@crvernon](https://github.com/crvernon), thanks a lot it needs to be a brief review of the paper, nothing like a monograph if that's ok, I'll add you as a reviewer

Hi @pachadotdev , @danielskatz is your associate editor-in-chief on this submission as this now falls under his CSISM track.  I'll transfer your inquiry to him.
