### 2025-10-30 #8287 @danielskatz
@realpaver - I'm going to take the editor over
