### 2025-10-31 #8287 @cvanaret
@danielskatz yes, I'm still up for it. I'm working on a paper revision due Nov 8, I'll write the review after that.
