### 2025-11-05 b1e968d
Update Readme.md to answer issue #2 about gurobi.pc file.
