### 2025-11-13 #9331 @dannys4
@danielskatz @realpaver I'm working through my review right now and, of course, I've tried installing the package. The package is nice in how modular it is---it allows you to switch between dependencies used as mathematical backends---but I seemingly cannot install GAOL, which is an optional dependency of this package, so I can't really review that part of the package (to make sure the provided package actually works with the backend).

Since the problem is out of scope of RealPaver, **is this a problem worth incorporating into my review or should I just work with the backend(s) I _can_ compile the project with**? I'd make an executive decision myself, but this would be good to know for future reviews of what's in-scope for C++ codes with complex dependency structures.
