### 2025-11-18 realpaver/realpaver#6 @victorapm
JOSS review - State of the field
Hi @realpaver @raphaelchenouard , thank you for the nice open-source work!

Regarding the JOSS paper https://github.com/openjournals/joss-reviews/issues/9331, would you be able to add a new section about "State of the field" and discuss how your software compares to other open-source solutions such as OR-Tools, Chuffed, and others?

Thanks!
