### 2025-11-18 realpaver/realpaver#7 @victorapm
JOSS review - Contributing instructions
Consider adding a file `CONTRIBUTING.md` with clear guidelines for:
 1. Contributing to the software
 2. Reporting issues or problems with the software
 3. Seeking support

Ref: https://github.com/openjournals/joss-reviews/issues/9331
