### 2025-11-20 realpaver/realpaver#6 @raphaelchenouard
Hi @victorapm, that you for your help in improving RealPaver and the JOSS paper.
The main software that is comparable to RealPaver is IBEX and that is already mentioned just before the statement of need.
OR-tools and Chuffed are solvers handling discrete problems, whereas RealPaver (and IBEX) solves continuous and mixed integer non-linear problems. To our knowledge, most other existing solvers mixed problems like Couenne or Baron focus on optimization problems only, so that they cannot be compared to RealPaver.
We don't think adding a section to the statement of the field is required, but if you do not agree, we'll add a paragraph like the text above.
