### 2025-11-20 realpaver/realpaver#9 @raphaelchenouard
Thanks for this issue, we are working on this issue with GAOL library main developer to find a solution. As @dannys4 mentioned this, we may work on a fork of GAOL with meson build system.
We also have a Dockerfile, and we'll add it to the repository to ease the testing of the library in a container.
