### 2025-11-20 fe0fb6f
Modifies the tests of var3B, polytope hull, and LP solver
