### 2025-11-20 realpaver/realpaver#6 @victorapm
Hi @raphaelchenouard, thanks for the added context and apologies for missing the earlier mention of IBEX!

The explanation you just provided regarding Couenne and Baron (focusing on optimization vs. finding solution sets) sounds good. It provides exactly the kind of context a reader needs to understand where RealPaver fits in the ecosystem, and I strongly suggest adding that distinction to the paper.

Regarding IBEX: while the text currently implies they produce similar results, a proper "State of the field" description should help a potential user decide which tool to use. For example, does RealPaver offer a different API design, distinct constraint support, or a more modular architecture compared to IBEX? In essence, how do they differ at a high-level?

Could you please add the paragraph you proposed, but also explicitly **articulate** the functional differences or distinct design goals of RealPaver relative to IBEX? This would improve the paper toward the "State of the field" requirement in JOSS.

cc @danielskatz
