### 2025-11-20 fb667a0
Add CONTRIBUTING.md to answer issue #7
