+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+# Contributing to RealPaver
+
+Thank you for your interest in RealPaver!
+
+## Report issues
+
+If you encounter a bug, have a feature request, or find something unclear in the documentation, please open a [GitHub Issue](https://github.com/realpaver/realpaver/issues).
+
+## How to contribute
+
+Contributions are welcome:
+
+- Code Contributions: to correct bugs, improve the codebase, enhance performance, or implement new mechanisms, you can submit a [GitHub Pull Request](https://github.com/realpaver/realpaver/pulls)
+- Documentation Improvements: to add information that is missing or to improve unclear explanations, you are welcome to submit a [GitHub Pull Request](https://github.com/realpaver/realpaver/pulls)
+
+## RealPaver discussions or support
+
+If you are seeking support or want to discuss about RealPaver itself, feel free to participate in existing discussions or create a new one via [GitHub Discussions](https://github.com/realpaver/realpaver/discussions).
