### 2025-11-20 realpaver/realpaver#7 @raphaelchenouard
Thanks, the latest commit (https://github.com/realpaver/realpaver/commit/fb667a08c2d944451011d290707cf6f089060bf4) should answer you demand.
