### 2025-11-21 d9d4d18
Adds an example joss_test to test chebyshevCos
