### 2025-11-25 e9904c3
Identifies inum64 as a dependency
