+++ src/realpaver/ContractorPolytope.cpp
@@ -170,10 +170,15 @@ Proof ContractorPolytope::run(LPSolver &solver, IntervalBox &B, int *lb, int &nl
 
       LOG_LOW("LP solved: " << sense << " " << v.getName());
 
+      DEBUG("\n\nLP solved: " << sense << " " << v.getName());
+      solver.print(std::cerr);
+      std::cerr << std::endl;
+
       // optimization
       status = (status != LPStatus::Optimal) ? solver.optimize() : solver.reoptimize();
 
       LOG_LOW("Status : " << status);
+      DEBUG("Status : " << status);
 
       switch (status)
       {
@@ -182,6 +187,7 @@ Proof ContractorPolytope::run(LPSolver &solver, IntervalBox &B, int *lb, int &nl
          LOG_LOW("Certified cost: " << bnd);
          LOG_LOW("Primal: " << solver.primalSolution());
 
+         DEBUG("bnd = " << bnd);
          Interval x = B.get(v);
          if (sense == LPSense::Min)
          {
