+++ meson_options.txt
@@ -15,7 +15,7 @@
 #-------------------------------------------------------------------------------
 
 option('LP_LIB', type : 'string', value : 'Highs', description : 'Linear Programming library (Highs, Clp, Gurobi)')
-option('INTERVAL_LIB', type : 'string', value : 'Gaol', description : 'Interval arithmetic library (Gaol)')
+option('INTERVAL_LIB', type : 'string', value : 'Gaol', description : 'Interval arithmetic library (Gaol, Inum64)')
 option('ASSERT', type : 'boolean', value : false, description : 'Activate assertions')
 option('LOG', type : 'boolean', value : false, description : 'Activate file logging')
 option('GENERATE_DOC', type : 'boolean', value: false, description : 'Activate the generation of the documentation (Doxygen)')
