### 2025-11-25 dd99204
Update of paper.md to add a state of the field.
