### 2025-11-25 realpaver/realpaver#6 @raphaelchenouard
The commit dd992047158870b46542e2ccab5064e23f59564f add a state of the field as you asked.
