### 2025-11-25 f351683
Fixes a bug in the interface of the LP solver Highs
