### 2025-11-25 7538b90
ASSERT, DEBUG and LOG constants have been rename to add the prefix REALPAVER_.
