+++ src/realpaver/RealPoint.cpp
@@ -18,8 +18,8 @@
  * @date   2024-4-11
  */
 
-#include "realpaver/RealPoint.hpp"
 #include "realpaver/AssertDebug.hpp"
+#include "realpaver/RealPoint.hpp"
 
 namespace realpaver {
 
@@ -27,22 +27,23 @@ RealPoint::RealPoint(Scope scop, double a)
     : RealVector(scop.size(), a)
     , scop_(scop)
 {
-   ASSERT(!scop.isEmpty(), "Empty scope used to create a real point");
+   REALPAVER_ASSERT(!scop.isEmpty(), "Empty scope used to create a real point");
 }
 
 RealPoint::RealPoint(Scope scop, const RealVector &X)
     : RealVector(X)
     , scop_(scop)
 {
-   ASSERT(!scop.isEmpty(), "Empty scope used to create a real point");
-   ASSERT(scop.size() == X.size(), "Bad initialization of a real point");
+   REALPAVER_ASSERT(!scop.isEmpty(), "Empty scope used to create a real point");
+   REALPAVER_ASSERT(scop.size() == X.size(), "Bad initialization of a real point");
 }
 
 RealPoint::RealPoint(const RealPoint &pt, Scope scop)
     : RealVector(scop.size())
     , scop_(scop)
 {
-   ASSERT(pt.scope().contains(scop), "Bad scope used to project a point in a sub-space");
+   REALPAVER_ASSERT(pt.scope().contains(scop),
+                    "Bad scope used to project a point in a sub-space");
 
    for (const auto &v : scop)
       set(v, pt.get(v));
@@ -55,14 +56,14 @@ Scope RealPoint::scope() const
 
 double RealPoint::get(const Variable &v) const
 {
-   ASSERT(scop_.contains(v), "Bad access in a real point @ " << v.getName());
+   REALPAVER_ASSERT(scop_.contains(v), "Bad access in a real point @ " << v.getName());
 
    return operator[](scop_.index(v));
 }
 
 void RealPoint::set(const Variable &v, double a)
 {
-   ASSERT(scop_.contains(v), "Bad access in a real point @ " << v.getName());
+   REALPAVER_ASSERT(scop_.contains(v), "Bad access in a real point @ " << v.getName());
 
    operator[](scop_.index(v)) = a;
 }
@@ -75,7 +76,8 @@ void RealPoint::setOnScope(const RealPoint &pt, const Scope &scop)
 
 RealPoint RealPoint::subPoint(const Scope &scop) const
 {
-   ASSERT(scop_.contains(scop), "Bad scope used to create a sub-point " << scop);
+   REALPAVER_ASSERT(scop_.contains(scop),
+                    "Bad scope used to create a sub-point " << scop);
 
    RealPoint pt(scop);
    for (const auto &v : scop)
