### 2025-11-25 realpaver/realpaver#9 @raphaelchenouard
@victorapm and @dannys4
I updated my fork of GAOL to add meson scripts in the master branch to compile it under MacOS with ARM SoCs:  https://github.com/raphaelchenouard/GAOL

I have to check how to handle the installation of crlibm or apmathlib from this fork since new platforms (like MacOS ARM) may not be supported by these libraries, since they rely on autotools. However, if one these are already installed on your system, it should work.
