### 2025-11-25 30b8897
Update README.md to add a link to a GAOL fork to use meson build system to compile and install it.
