### 2025-11-26 realpaver/realpaver#9 @victorapm
Thank you, @raphaelchenouard

I'm having issues building `crlibm` and `apmathlib` on `arm64`. See below:

```
~/projects/mathlib-2.1.1  % ./configure --prefix=$(pwd)/install --build=arm

./configure: line 428: test: this: integer expression expected
configure: error: system
./configure: line 313: return: this: numeric argument required
./configure: line 323: exit: this: numeric argument required

```

```
~/projects/crlibm % ./configure && make all

clang -O0 -std=gnu99  -g -O2   -o crlibm_testval test_val.o test_common.o ../libcrlibm.a -lm
ld: archive member '/' not a mach-o file in 'projects/crlibm/libcrlibm.a'
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[2]: *** [crlibm_testval] Error 1
make[1]: *** [all-recursive] Error 1
make: *** [all] Error 2
```

If resolving these issues proves too time-consuming, another option is simply to state that `realpaver` is currently not supported on `arm64`
