### 2025-11-26 realpaver/realpaver#9 @raphaelchenouard
You're right there is no official repo for apmathlib and the one for crlibm does not seem maintained anymore.
I found this one from which I was able to compile with default `configure`and `make` command (i.e. without any other args): https://github.com/taschini/pycrlibm (it is mentionned in the readme of my GAOL fork).

As suggested by @dannys4, I also pushed a few dockerfiles in a docker folder at the root of RealPaver repository.
