### 2025-11-26 realpaver/realpaver#9 @raphaelchenouard
In fact I've made:

```bash
git clone https://github.com/taschini/pycrlibm
cd pycrlibm
cd crlibm
./configure
make
```

This way, I was able to compile it. If you want to install to a specific location, you can do for instance:
```bash
./configure --prefix=/opt/homebrew/Cellar/crlibm/1.0
make
make install
```

If you do so and have homebrew, then:

```bash
brew link crlibm
```
