### 2025-11-26 realpaver/realpaver#9 @victorapm
This is the output I'm getting:

```
~/projects  % git clone https://github.com/taschini/pycrlibm
Cloning into 'pycrlibm'...
remote: Enumerating objects: 403, done.
remote: Total 403 (delta 0), reused 2 (delta 0), pack-reused 400 (from 1)
Resolving deltas: 100% (120/120), done.

~/projects  % cd pycrlibm

~/projects/pycrlibm (master) % cd crlibm

~/projects/pycrlibm/crlibm (master) % ./configure
checking for a BSD-compatible install... /opt/homebrew/bin/ginstall -c
checking whether build environment is sane... yes
checking for a thread-safe mkdir -p... /opt/homebrew/bin/gmkdir -p
checking for gawk... no
checking for mawk... no
checking for nawk... no
checking for awk... awk
checking whether make sets $(MAKE)... yes
checking build system type... mkdir: /var/folders/zz/zyxvpxvq6csfxvn_n0000000000000/T//cg52011-11129: Permission denied
mkdir: /var/folders/zz/zyxvpxvq6csfxvn_n0000000000000/T//cg-52011: Permission denied
config.guess: cannot create a temporary directory in /var/folders/zz/zyxvpxvq6csfxvn_n0000000000000/T/
configure: error: cannot guess build type; you must specify one
```

So `configure` doesn't recognize the system. It works if I tell it's arm:

```
% ./configure --build=arm
checking for a BSD-compatible install... /opt/homebrew/bin/ginstall -c
checking whether build environment is sane... yes
checking for a thread-safe mkdir -p... /opt/homebrew/bin/gmkdir -p
checking for gawk... no
checking for mawk... no
checking for nawk... no
checking for awk... awk
checking whether make sets $(MAKE)... yes
checking build system type... arm-unknown-none
checking host system type... arm-unknown-none
checking whether make sets $(MAKE)... (cached) yes
checking for gcc... gcc
checking for C compiler default output file name... a.out
checking whether the C compiler works... yes
checking whether we are cross compiling... no
checking for suffix of executables...
checking for suffix of object files... o
checking whether we are using the GNU C compiler... yes
checking whether gcc accepts -g... yes
checking for gcc option to accept ISO C89... none needed
checking for style of include used by make... GNU
checking dependency style of gcc... gcc3
checking for ranlib... ranlib
checking whether ln -s works... yes
checking how to run the C preprocessor... gcc -E
checking for grep that handles long lines and -e... /usr/bin/grep
checking for egrep... /usr/bin/grep -E
checking for ANSI C header files... yes
checking for an ANSI C-conforming const... yes
checking for inline... inline
checking for sqrt... yes
checking for sys/types.h... yes
checking for sys/stat.h... yes
checking for stdlib.h... yes
checking for string.h... yes
checking for memory.h... yes
checking for strings.h... yes
checking for inttypes.h... yes
checking for stdint.h... yes
checking for unistd.h... yes
checking fenv.h usability... yes
checking fenv.h presence... yes
checking for fenv.h... yes
checking float.h usability... yes
checking float.h presence... yes
checking for float.h... yes
checking fpu_control.h usability... no
checking fpu_control.h presence... no
checking for fpu_control.h... no
checking for log in -lm... yes
checking whether byte ordering is bigendian... no
checking for long double with more range or precision than double... no
checking size of int... 4
configure: creating ./config.status
config.status: creating Makefile
config.status: creating tests/Makefile
config.status: creating scs_lib/Makefile
config.status: creating crlibm_config.h
config.status: executing depfiles commands
```

But then, there's a compile issue (at the very end):

```
~/projects/pycrlibm/crlibm (master) % make all
/Library/Developer/CommandLineTools/usr/bin/make  all-recursive
Making all in scs_lib
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT scs_private.o -MD -MP -MF .deps/scs_private.Tpo -c -o scs_private.o scs_private.c
mv -f .deps/scs_private.Tpo .deps/scs_private.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT double2scs.o -MD -MP -MF .deps/double2scs.Tpo -c -o double2scs.o double2scs.c
mv -f .deps/double2scs.Tpo .deps/double2scs.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT print_scs.o -MD -MP -MF .deps/print_scs.Tpo -c -o print_scs.o print_scs.c
mv -f .deps/print_scs.Tpo .deps/print_scs.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT division_scs.o -MD -MP -MF .deps/division_scs.Tpo -c -o division_scs.o division_scs.c
mv -f .deps/division_scs.Tpo .deps/division_scs.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT addition_scs.o -MD -MP -MF .deps/addition_scs.Tpo -c -o addition_scs.o addition_scs.c
mv -f .deps/addition_scs.Tpo .deps/addition_scs.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT multiplication_scs.o -MD -MP -MF .deps/multiplication_scs.Tpo -c -o multiplication_scs.o multiplication_scs.c
mv -f .deps/multiplication_scs.Tpo .deps/multiplication_scs.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT scs2double.o -MD -MP -MF .deps/scs2double.Tpo -c -o scs2double.o scs2double.c
mv -f .deps/scs2double.Tpo .deps/scs2double.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT zero_scs.o -MD -MP -MF .deps/zero_scs.Tpo -c -o zero_scs.o zero_scs.c
mv -f .deps/zero_scs.Tpo .deps/zero_scs.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..    -g -O2 -MT scs2mpfr.o -MD -MP -MF .deps/scs2mpfr.Tpo -c -o scs2mpfr.o scs2mpfr.c
mv -f .deps/scs2mpfr.Tpo .deps/scs2mpfr.Po
rm -f libscs.a
ar cru libscs.a scs_private.o double2scs.o print_scs.o division_scs.o addition_scs.o multiplication_scs.o scs2double.o zero_scs.o scs2mpfr.o
ar: `u' modifier ignored since `D' is the default (see `U')
ranlib libscs.a
Making all in .
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT crlibm_private.o -MD -MP -MF .deps/crlibm_private.Tpo -c -o crlibm_private.o crlibm_private.c
mv -f .deps/crlibm_private.Tpo .deps/crlibm_private.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT triple-double.o -MD -MP -MF .deps/triple-double.Tpo -c -o triple-double.o triple-double.c
mv -f .deps/triple-double.Tpo .deps/triple-double.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT exp-td.o -MD -MP -MF .deps/exp-td.Tpo -c -o exp-td.o exp-td.c
exp-td.c:112:53: warning: unused variable 't4db2' [-Wunused-variable]
  112 |   db_number shiftedXMultdb, twoPowerMdb, xdb, t4db, t4db2, polyTblhdb, resdb;
      |                                                     ^~~~~
exp-td.c:391:53: warning: unused variable 't4db2' [-Wunused-variable]
  391 |   db_number shiftedXMultdb, twoPowerMdb, xdb, t4db, t4db2, resdb;
      |                                                     ^~~~~
exp-td.c:684:53: warning: unused variable 't4db2' [-Wunused-variable]
  684 |   db_number shiftedXMultdb, twoPowerMdb, xdb, t4db, t4db2, resdb;
      |                                                     ^~~~~
3 warnings generated.
mv -f .deps/exp-td.Tpo .deps/exp-td.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT exp-td-standalone.o -MD -MP -MF .deps/exp-td-standalone.Tpo -c -o exp-td-standalone.o exp-td-standalone.c
mv -f .deps/exp-td-standalone.Tpo .deps/exp-td-standalone.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT expm1-standalone.o -MD -MP -MF .deps/expm1-standalone.Tpo -c -o expm1-standalone.o expm1-standalone.c
mv -f .deps/expm1-standalone.Tpo .deps/expm1-standalone.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT expm1.o -MD -MP -MF .deps/expm1.Tpo -c -o expm1.o expm1.c
mv -f .deps/expm1.Tpo .deps/expm1.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT log.o -MD -MP -MF .deps/log.Tpo -c -o log.o log.c
mv -f .deps/log.Tpo .deps/log.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT log1p.o -MD -MP -MF .deps/log1p.Tpo -c -o log1p.o log1p.c
log1p.c:208:8: warning: variable 'E' is used uninitialized whenever 'if' condition is true [-Wsometimes-uninitialized]
  208 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
log1p.c:419:7: note: uninitialized use occurs here
  419 |    if(E==0)
      |       ^
log1p.c:208:4: note: remove the 'if' if its condition is always false
  208 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  209 |      /* Use the polynomial p(zh + zl) approximating log(1+zh+zl) directly
      |      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  210 |         Set E and index to values that read 0.0 in the accurate phase.
      |         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  211 |      */
      |      ~~
  212 |      logih = 0.0;
      |      ~~~~~~~~~~~~
  213 |      logim = 0.0;
      |      ~~~~~~~~~~~~
  214 |      index = 0;
      |      ~~~~~~~~~~
  215 |      ed = 0.0;
      |      ~~~~~~~~~
  216 |      index = 0;
      |      ~~~~~~~~~~
  217 |      zh = x;
      |      ~~~~~~~
  218 |      zm = 0.0;
      |      ~~~~~~~~~
  219 |      zl = 0.0;
      |      ~~~~~~~~~
  220 |    } else {
      |    ~~~~~~
log1p.c:171:9: note: initialize the variable 'E' to silence this warning
  171 |    int E, index;
      |         ^
      |          = 0
log1p.c:496:8: warning: variable 'E' is used uninitialized whenever 'if' condition is true [-Wsometimes-uninitialized]
  496 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
log1p.c:707:7: note: uninitialized use occurs here
  707 |    if(E==0)
      |       ^
log1p.c:496:4: note: remove the 'if' if its condition is always false
  496 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  497 |      /* Use the polynomial p(zh + zl) approximating log(1+zh+zl) directly
      |      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  498 |         Set E and index to values that read 0.0 in the accurate phase.
      |         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  499 |      */
      |      ~~
  500 |      logih = 0.0;
      |      ~~~~~~~~~~~~
  501 |      logim = 0.0;
      |      ~~~~~~~~~~~~
  502 |      index = 0;
      |      ~~~~~~~~~~
  503 |      ed = 0.0;
      |      ~~~~~~~~~
  504 |      index = 0;
      |      ~~~~~~~~~~
  505 |      zh = x;
      |      ~~~~~~~
  506 |      zm = 0.0;
      |      ~~~~~~~~~
  507 |      zl = 0.0;
      |      ~~~~~~~~~
  508 |    } else {
      |    ~~~~~~
log1p.c:450:9: note: initialize the variable 'E' to silence this warning
  450 |    int E, index;
      |         ^
      |          = 0
log1p.c:783:8: warning: variable 'E' is used uninitialized whenever 'if' condition is true [-Wsometimes-uninitialized]
  783 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
log1p.c:994:7: note: uninitialized use occurs here
  994 |    if(E==0)
      |       ^
log1p.c:783:4: note: remove the 'if' if its condition is always false
  783 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  784 |      /* Use the polynomial p(zh + zl) approximating log(1+zh+zl) directly
      |      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  785 |         Set E and index to values that read 0.0 in the accurate phase.
      |         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  786 |      */
      |      ~~
  787 |      logih = 0.0;
      |      ~~~~~~~~~~~~
  788 |      logim = 0.0;
      |      ~~~~~~~~~~~~
  789 |      index = 0;
      |      ~~~~~~~~~~
  790 |      ed = 0.0;
      |      ~~~~~~~~~
  791 |      index = 0;
      |      ~~~~~~~~~~
  792 |      zh = x;
      |      ~~~~~~~
  793 |      zm = 0.0;
      |      ~~~~~~~~~
  794 |      zl = 0.0;
      |      ~~~~~~~~~
  795 |    } else {
      |    ~~~~~~
log1p.c:729:9: note: initialize the variable 'E' to silence this warning
  729 |    int E, index;
      |         ^
      |          = 0
log1p.c:1070:8: warning: variable 'E' is used uninitialized whenever 'if' condition is true [-Wsometimes-uninitialized]
 1070 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
log1p.c:1281:7: note: uninitialized use occurs here
 1281 |    if(E==0)
      |       ^
log1p.c:1070:4: note: remove the 'if' if its condition is always false
 1070 |    if ((xdb.i[HI] & 0x7fffffff) < 0x3f700000) {
      |    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 1071 |      /* Use the polynomial p(zh + zl) approximating log(1+zh+zl) directly
      |      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 1072 |         Set E and index to values that read 0.0 in the accurate phase.
      |         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 1073 |      */
      |      ~~
 1074 |      logih = 0.0;
      |      ~~~~~~~~~~~~
 1075 |      logim = 0.0;
      |      ~~~~~~~~~~~~
 1076 |      index = 0;
      |      ~~~~~~~~~~
 1077 |      ed = 0.0;
      |      ~~~~~~~~~
 1078 |      index = 0;
      |      ~~~~~~~~~~
 1079 |      zh = x;
      |      ~~~~~~~
 1080 |      zm = 0.0;
      |      ~~~~~~~~~
 1081 |      zl = 0.0;
      |      ~~~~~~~~~
 1082 |    } else {
      |    ~~~~~~
log1p.c:1016:9: note: initialize the variable 'E' to silence this warning
 1016 |    int E, index;
      |         ^
      |          = 0
4 warnings generated.
mv -f .deps/log1p.Tpo .deps/log1p.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT rem_pio2_accurate.o -MD -MP -MF .deps/rem_pio2_accurate.Tpo -c -o rem_pio2_accurate.o rem_pio2_accurate.c
rem_pio2_accurate.c:93:14: warning: equality comparison with extraneous parentheses [-Wparentheses-equality]
   93 |   if ((X_IND == -1)){
      |        ~~~~~~^~~~~
rem_pio2_accurate.c:93:14: note: remove extraneous parentheses around the comparison to silence this warning
   93 |   if ((X_IND == -1)){
      |       ~      ^    ~
rem_pio2_accurate.c:93:14: note: use '=' to turn this equality comparison into an assignment
   93 |   if ((X_IND == -1)){
      |              ^~
      |              =
1 warning generated.
mv -f .deps/rem_pio2_accurate.Tpo .deps/rem_pio2_accurate.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT trigo_fast.o -MD -MP -MF .deps/trigo_fast.Tpo -c -o trigo_fast.o trigo_fast.c
trigo_fast.c:175:14: warning: equality comparison with extraneous parentheses [-Wparentheses-equality]
  175 |   if ((X_IND == -2)){
      |        ~~~~~~^~~~~
trigo_fast.c:175:14: note: remove extraneous parentheses around the comparison to silence this warning
  175 |   if ((X_IND == -2)){
      |       ~      ^    ~
trigo_fast.c:175:14: note: use '=' to turn this equality comparison into an assignment
  175 |   if ((X_IND == -2)){
      |              ^~
      |              =
trigo_fast.c:407:15: warning: equality comparison with extraneous parentheses [-Wparentheses-equality]
  407 |     if((index == 0)) {
      |         ~~~~~~^~~~
trigo_fast.c:407:15: note: remove extraneous parentheses around the comparison to silence this warning
  407 |     if((index == 0)) {
      |        ~      ^   ~
trigo_fast.c:407:15: note: use '=' to turn this equality comparison into an assignment
  407 |     if((index == 0)) {
      |               ^~
      |               =
2 warnings generated.
mv -f .deps/trigo_fast.Tpo .deps/trigo_fast.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT trigo_accurate.o -MD -MP -MF .deps/trigo_accurate.Tpo -c -o trigo_accurate.o trigo_accurate.c
mv -f .deps/trigo_accurate.Tpo .deps/trigo_accurate.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT trigpi.o -MD -MP -MF .deps/trigpi.Tpo -c -o trigpi.o trigpi.c
mv -f .deps/trigpi.Tpo .deps/trigpi.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT asincos.o -MD -MP -MF .deps/asincos.Tpo -c -o asincos.o asincos.c
asincos.c:3450:29: warning: variable 'deltal' set but not used [-Wunused-but-set-variable]
 3450 |   double deltatemp, deltah, deltal;
      |                             ^
asincos.c:3804:29: warning: variable 'deltal' set but not used [-Wunused-but-set-variable]
 3804 |   double deltatemp, deltah, deltal;
      |                             ^
asincos.c:4157:29: warning: variable 'deltal' set but not used [-Wunused-but-set-variable]
 4157 |   double deltatemp, deltah, deltal;
      |                             ^
3 warnings generated.
mv -f .deps/asincos.Tpo .deps/asincos.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT pow.o -MD -MP -MF .deps/pow.Tpo -c -o pow.o pow.c
mv -f .deps/pow.Tpo .deps/pow.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT atan_fast.o -MD -MP -MF .deps/atan_fast.Tpo -c -o atan_fast.o atan_fast.c
mv -f .deps/atan_fast.Tpo .deps/atan_fast.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT atan_accurate.o -MD -MP -MF .deps/atan_accurate.Tpo -c -o atan_accurate.o atan_accurate.c
mv -f .deps/atan_accurate.Tpo .deps/atan_accurate.Po
gcc -DHAVE_CONFIG_H -I.    -std=c99 -Wall -O3 -g -O2 -MT csh_fast.o -MD -MP -MF .deps/csh_fast.Tpo -c -o csh_fast.o csh_fast.c
mv -f .deps/csh_fast.Tpo .deps/csh_fast.Po
rm -f libcrlibm.a
ar cru libcrlibm.a crlibm_private.o triple-double.o exp-td.o exp-td-standalone.o expm1-standalone.o expm1.o log.o log1p.o rem_pio2_accurate.o trigo_fast.o trigo_accurate.o trigpi.o asincos.o pow.o atan_fast.o atan_accurate.o csh_fast.o scs_lib/scs_private.o scs_lib/addition_scs.o scs_lib/division_scs.o scs_lib/print_scs.o scs_lib/double2scs.o scs_lib/zero_scs.o scs_lib/multiplication_scs.o scs_lib/scs2double.o
ar: `u' modifier ignored since `D' is the default (see `U')
ranlib libcrlibm.a
Making all in tests
gcc -DHAVE_CONFIG_H -I. -I.. -I..   -O0 -std=gnu99  -g -O2 -MT test_val.o -MD -MP -MF .deps/test_val.Tpo -c -o test_val.o test_val.c
test_val.c:156:48: warning: passing arguments to a function without a prototype is deprecated in all versions of C and is not supported in C2x [-Wdeprecated-non-prototype]
  156 |     if (nbarg==1) res_crlibm.d = testfun_crlibm(input.d);
      |                                                ^
test_val.c:157:48: warning: passing arguments to a function without a prototype is deprecated in all versions of C and is not supported in C2x [-Wdeprecated-non-prototype]
  157 |     if (nbarg==2) res_crlibm.d = testfun_crlibm(input.d, input2.d);
      |                                                ^
test_val.c:228:44: warning: passing arguments to a function without a prototype is deprecated in all versions of C and is not supported in C2x [-Wdeprecated-non-prototype]
  228 |     if (nbarg==1) res_libm.d = testfun_libm(input.d);
      |                                            ^
test_val.c:229:44: warning: passing arguments to a function without a prototype is deprecated in all versions of C and is not supported in C2x [-Wdeprecated-non-prototype]
  229 |     if (nbarg==2) res_libm.d = testfun_libm(input.d, input2.d);
      |                                            ^
4 warnings generated.
mv -f .deps/test_val.Tpo .deps/test_val.Po
gcc -DHAVE_CONFIG_H -I. -I.. -I..   -O0 -std=gnu99  -g -O2 -MT test_common.o -MD -MP -MF .deps/test_common.Tpo -c -o test_common.o test_common.c
mv -f .deps/test_common.Tpo .deps/test_common.Po
gcc -O0 -std=gnu99  -g -O2   -o crlibm_testval test_val.o test_common.o ../libcrlibm.a -lm
ld: archive member '/' not a mach-o file in '~/projects/pycrlibm/crlibm/libcrlibm.a'
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[2]: *** [crlibm_testval] Error 1
make[1]: *** [all-recursive] Error 1
make: *** [all] Error 2

```
