### 2025-11-26 realpaver/realpaver#9 @dannys4
FWIW, I couldn't get the pycrlibm version to work, but I got `crlibm` to compile on MacOS using the tarball provided here: https://github.com/JuliaIntervals/CRlibm.jl/raw/368ccdebd57e7a7cc96d47ab6b08e7a597495ff2/deps/src/crlibm-1.0beta4.tar.gz
