### 2025-11-26 realpaver/realpaver#9 @victorapm
Thanks, @dannys4! I'm still seeing the same build error

```
clang -DHAVE_CONFIG_H -I. -I.. -I..   -O0 -std=gnu99  -g -O2 -MT test_common.o -MD -MP -MF .deps/test_common.Tpo -c -o test_common.o test_common.c
mv -f .deps/test_common.Tpo .deps/test_common.Po
clang -O0 -std=gnu99  -g -O2   -o crlibm_testval test_val.o test_common.o ../libcrlibm.a -lm
ld: archive member '/' not a mach-o file in '/Users/paludettomag1/projects/crlibm-1.0beta4/libcrlibm.a'
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[2]: *** [crlibm_testval] Error 1
make[1]: *** [all-recursive] Error 1
make: *** [all] Error 2
```
