### 2025-11-26 realpaver/realpaver#9 @victorapm
Looking at this more carefully... `libcrlibm.a` is being built successfully, the compile issue above is about crlibm tests and therefore not a blocking issue. @raphaelchenouard maybe you could a pointer to the installation instructions about this
