### 2025-11-26 realpaver/realpaver#9 @dannys4
@victorapm Sorry my suggestion wasn't of help

@raphaelchenouard Looking through the meson code, I had to make a few adjustments. First, I added the following file to `/my/install/path/CRLIBM/lib/pkgconfig/crlibm.pc`:

```pkgconfig
prefix=/my/install/path/CRLIBM
includedir=${prefix}/include
libdir=${prefix}/lib

Name: crlibm
Description: crlibm: crlibm
Version: 1.0.0
Cflags: -I${includedir}
Libs: -L${libdir} -lcrlibm
```

Then, I changed the Meson and pushed it up to a branch `deps` in my fork here: https://github.com/dannys4/GAOL/tree/deps

Finally, GAOL seemed to compile and the example seemed to run.
