### 2025-11-26 realpaver/realpaver#9 @dannys4
I also found that, before I could compile realpaver with GAOL, I had to also bump the C++ std to 14, since realpaver uses `std::make_unique` (introduced in c++14) (unrelated to CRLIBM)
