### 2025-11-26 realpaver/realpaver#13 @dannys4
Example `parsing` not working
It's my understanding that the file `examples/parsing.cpp` provides a file that just takes in a `param.txt` text doc and prints out the parameters within this text document. When I run `./build/examples/parsing param.txt` for the given parameters file (pasted below), I get a parse error `../examples/parsing.cpp(29): Error: Parse error: l2: syntax error`. Should I worry about this? it's not really documented, but I figured that I should be able to run all the examples.

```
# in seconds
TIME_LIMIT=1000

# No postprocessing
CERTIFICATION=NO

# Display traces in terminal
TRACE=YES

# Absolute tolerance on the width of intervals
VAR_ABS_TOL=1e-10

# Limit on the depth of nodes of search tree
DEPTH_LIMIT=1000

# Limit on the number of nodes generated by the solving algorithm
NODE_LIMIT=1000000

#

```
