### 2025-11-26 #9331 @dannys4
Okay, pending the two small points in my review-in-progress regarding the article itself, I'm pretty happy to call my review more-or-less complete. I have a few small PRs up regarding docs (https://github.com/realpaver/realpaver/pull/12), issues working with GAOL (https://github.com/realpaver/realpaver/pull/11), and an issue with an (undocumented) example (https://github.com/realpaver/realpaver/issues/13); with the new provision of Dockerfiles---which are a huge boon for developers thinking of using the package!!!---most of my concerns have been alleviated.

I'd like to extend appreciation for @realpaver's patience here, since this has been somewhat more involved that the reviews I've previously done with JOSS, and their flexibility/reception of feedback for the build systems.
