### 2025-11-27 realpaver/realpaver#13 @raphaelchenouard
You're right, it's not clearly documented and you've misunderstood its purpose. The parser aims at parsing RealPaver problem files (`.rp`) and not parameter files. Then, you can call it this way:

```bash
./build/examples/parsing ./examples/test.rp
```

In parsing.cpp, to build the parser object it requires a Params object and in the code, a default one is created (l.23), then it is passed to the parser constructor (l.24). Line 27 the parser is asked to parse the filename given as argument of the command line, and it fills the Problem object with read constants, variables, aliases and constraints.
