### 2025-11-27 realpaver/realpaver#9 @raphaelchenouard
Normally, the c++ version is defined as c++17 (l.18 of meson.build in the project definition), so you shouldn't face this issue.
