### 2025-11-27 realpaver/realpaver#9 @raphaelchenouard
I'm able to reproduce your error about `std::make_unique` until I use the dependency function of meson for GAOL. In this case, the compilation command uses at the same time `-std=c++17`and `-std=c++11` and the compiler takes only c++11 instead of c++17.
I think this is a meson issue and a compiler one that should take the most restrictive option.
