### 2025-12-09 realpaver/realpaver#9 @raphaelchenouard
I updated my GAOL repository to remove the extra_cflags argument in the pkgconfig generation task, that is adding this useless c++11 flag then generating these errors.
Now everything compile with all tests passing, on my setup for all configurations (macos with clp/highs and gaol/inum64 or any of the 4 docker corresponding docker containers).
