### 2025-12-11 e19aa67
Fix useless warnings as described in #14
