+++ src/realpaver/LPSolverHighs.cpp
@@ -133,35 +133,37 @@ void LPSolver::setOptions()
    int maxsec = getMaxSeconds();
    int maxiter = getMaxIter();
    double tol = getFeasTol();
-   HighsStatus hs, ok = HighsStatus::kOk;
 
    // From the HiGHS' doc: HiGHS has efficient implementations of both the primal and dual
    // simplex methods, although the dual simplex solver is likely to be faster and is more
    // robust,
-   hs = simplex_->setOptionValue("solver", "simplex");
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   HighsStatus hs = simplex_->setOptionValue("solver", "simplex");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
    // From the HiGHS' doc: Strategy for simplex solver 0 => Choose; 1 => Dual (serial); 2
    // => Dual (SIP); 3 => Dual (PAMI); 4 => Primal
    hs = simplex_->setOptionValue("simplex_strategy", 1);
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
 
    hs = simplex_->setOptionValue("time_limit", maxsec);
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
 
    hs = simplex_->setOptionValue("simplex_iteration_limit", maxiter);
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
 
    hs = simplex_->setOptionValue("log_to_console", false);
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
 
    hs = simplex_->setOptionValue("output_flag", false);
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
 
    hs = simplex_->setOptionValue("primal_feasibility_tolerance", tol);
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
 
    hs = simplex_->setOptionValue("dual_feasibility_tolerance", tol);
-   REALPAVER_ASSERT(hs == ok, "Bad option in Highs: ");
+   REALPAVER_ASSERT(hs == HighsStatus::kOk, "Bad option in Highs: ");
+
+   if (hs != HighsStatus::kOk)
+      std::cout << "Warning: unable to set some HiGHS options" << std::endl;
 }
 
 LPStatus LPSolver::run()
@@ -247,7 +249,7 @@ bool LPSolver::infeasibleRay(RealVector &ray) const
    if ((status != LPStatus::Infeasible) && (status != LPStatus::InfeasibleOrUnbounded))
       return false;
 
-   int p = getNbLinVars() + getNbLinCtrs();
+   size_t p = getNbLinVars() + getNbLinCtrs();
    if (ray.size() != p)
       ray.resize(p);
 
