### 2025-12-15 642d5f9
Update README.md to warn the user about math library dependencies.
