### 2025-12-15 b8df867
Update paper.md to improve the state of needs as asked in #15
