### 2025-12-16 #9331 @realpaver
Thanks @danielskatz, I hope that most of the issues and concerns has been addressed by latest commits.

The comparison with ibex asked by @cvanaret is legitimate, but we don't want to oppose the 2 libraries, since we have participated also in ibex.
With realpaver, we mainly focus here on MINLP with specific constraints dedicated to handle engineering problems like piecewise and table constraints that does not exists in ibex as briefly explained in the state of the field. If you think it is mandatory, we can make a small benchmark on well-known NLP problems, but this is not our target and we don't want that it is the main point readers keep in mind after reading the paper.
