### 2025-12-16 22cae1e
Update paper.md in accordance with https://github.com/openjournals/joss-reviews/issues/9331#issuecomment-3660613823
