### 2026-01-05 0b4ece4
minor changes in paper.md
