+++ paper/paper.md
@@ -30,20 +30,20 @@ Constraint Programming (CP) is a paradigm for solving constraint satisfaction an
 
 RealPaver is a C++ library for CP over numeric or mixed discrete-continuous domains. Constraint Satisfaction Problems (CSPs) can be described either in C++ with the API or in a text file using the syntax of RealPaver specific language.  Then, they can be solved using the C++ API or using the CSP solver from the command line. The CSP solver is pre-configured, but various parameters can be modified in another text file.
 
-With respect to the first version of the software developed twenty years ago [@granvilliers2006], this new library incorporates new types of variables and constraints, new algorithms, a clean object-oriented architecture, the management of parameters, Meson Build as build engine [@meson], an interface with third-party softwares and a C++ API.
+With respect to the first version of the software developed twenty years ago [@granvilliers2006], this new library incorporates new types of variables and constraints, new algorithms, a clean object-oriented architecture, the management of parameters, Meson Build as build engine [@meson], an interface with third-party softwares, and a C++ API.
 
 
 # Statement of need
 
-CP associates a rich modeling language with powerful solving techniques. The main algorithm behind the RealPaver solver is a branch-and-prune (B&P) that implements a complete search to find all the solutions of a given problem [@pvh1997; @chabert2009]. The branching component separates a problem into sub-problems easier to solve. The pruning or contracting component aims at reducing the region delimited by a sub-problem. RealPaver relies on the GAOL interval library [@GAOL] to ensure rigorous computations.
+CP associates a rich modeling language with powerful solving techniques. The main algorithm behind the RealPaver solver is branch-and-prune (B&P), which implements a complete search to find all the solutions of a given problem [@pvh1997; @chabert2009]. The branching component separates a problem into easier-to-solve sub-problems. The pruning or contracting component aims at reducing the region delimited by a sub-problem. RealPaver relies on the GAOL interval library [@GAOL] to ensure rigorous computations.
 
-This technology has been applied with success in many fields of engineering like automatic control [@jaulin2001], preliminary design [@yvars2021] and robotics [@merlet2004].
+This technology has been applied with success in many fields of engineering like automatic control [@jaulin2001], preliminary design [@yvars2021], and robotics [@merlet2004].
 
 The set of solutions of a numerical CSP or a mixed discrete-continuous problem can be rigorously bounded, even in the presence of roundoff errors. All the solutions of continuous nonlinear equations or inequations can be automatically separated and rigorously bounded. The B\&P algorithm with interval analysis [@moore2009] can provide proof of infeasibility or proof of existence of solutions.
 
 # State of the field
 
-Various solvers on discrete domains exist in CP, like OR-tools and Chuffed, but they are not able to handle continuous variables. Ibex [@ibex] library like RealPaver are based on a similar branch-and-contract framework, but RealPaver proposes some new kinds of constraints like table constraints, conditional constraints and piecewise constraints, and the native definition of discrete variables (integer, enumeration of values). For pure  continuous problems, RealPaver achieves performances equivalent to Ibex. Regarding other mixed integer non-linear solvers like Couenne or Baron, they only solve optimization problems. These solvers are not able to compute several solutions for CSPs, even by adding a constant objective.
[…]
