### 2026-01-16 #9331 @realpaver
I've followed my institution recommandation and I put it on Software Heritage and on hal.science.
The software heritage id is:
swh:1:dir:e03f7408a774d4fe750eeb717e7d208cc76018b4;origin=https://github.com/realpaver/realpaver/archive/refs/tags/v1.1.1-joss.tar.gz;visit=swh:1:snp:d304f7e81097e427870d663e7b5f4b7ea2e0fce4 and the permalink is https://archive.softwareheritage.org/swh:1:dir:e03f7408a774d4fe750eeb717e7d208cc76018b4;origin=https://github.com/realpaver/realpaver/archive/refs/tags/v1.1.1-joss.tar.gz;visit=swh:1:snp:d304f7e81097e427870d663e7b5f4b7ea2e0fce4
It's waiting moderation validation on hal.science.
