### 2026-02-16 #9331 @realpaver
@danielskatz I was tired waiting for hal, so I put it on Zenodo: [10.5281/zenodo.18656438](https://doi.org/10.5281/zenodo.18656438).
The tag is [v1.1.1-joss2](https://github.com/realpaver/realpaver/releases/tag/v1.1.1-joss2)
Let me know if you need something else and sorry being late in answering.
