### 2026-02-16 #9331 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
