### 2026-02-16 #9331 @danielskatz
Congratulations to @realpaver (Raphael Chenouard) and co-author on your JOSS publication!!

And thanks to @cvanaret, @dannys4, and @victorapm for reviewing!
JOSS completely depends on volunteers; we couldn't do this without you
